Remove env contexts validation
Remove env context building and validation as app can never validate actions that set env variables dynamically (e.g., setup-dotnet). These are not defined in action.yml and would require to be extracted from the source code which can't be done reliably.

As the CLI prefers no false positives over false negatives to be trusted this feature is removed.